requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
psycopg2-binary
httpx
python-multipart
